step towards livewire
