+++ tests/Feature/Base/BaseSharingTestScenarios.php
@@ -17,7 +17,6 @@
 
 namespace Tests\Feature\Base;
 
-use App\Models\Configs;
 use App\SmartAlbums\OnThisDayAlbum;
 use App\SmartAlbums\RecentAlbum;
 use App\SmartAlbums\StarredAlbum;
@@ -57,9 +56,6 @@ use Tests\Feature\Constants\TestConstants;
  */
 abstract class BaseSharingTestScenarios extends BaseSharingTest
 {
-	/** @var bool the previously configured visibility for public photos */
-	protected bool $arePublicPhotosHidden;
-
 	/** @var string|null the ID of the current album 1 (if applicable by the test) */
 	protected ?string $albumID1;
 
@@ -84,9 +80,6 @@ abstract class BaseSharingTestScenarios extends BaseSharingTest
 	public function setUp(): void
 	{
 		parent::setUp();
-
-		$this->arePublicPhotosHidden = Configs::getValueAsBool(TestConstants::CONFIG_PUBLIC_HIDDEN);
-
 		// Reset all variables to ensure that a test implementation of the
 		// child class does not accidentally use a value which it should not.
 		$this->albumID1 = null;
@@ -100,7 +93,6 @@ abstract class BaseSharingTestScenarios extends BaseSharingTest
 
 	public function tearDown(): void
 	{
-		Configs::set(TestConstants::CONFIG_PUBLIC_HIDDEN, $this->arePublicPhotosHidden);
 		parent::tearDown();
 	}
 
@@ -139,7 +131,7 @@ abstract class BaseSharingTestScenarios extends BaseSharingTest
 	/**
 	 * Ensures that the user does not see the private photo.
 	 *
-	 * See {@link SharingTestScenariosAbstract::prepareUnsortedPublicAndPrivatePhoto()}
+	 * See {@link BaseSharingTestScenarios::prepareUnsortedPublicAndPrivatePhoto()}
 	 * for description of scenario.
 	 *
 	 * @return void
@@ -196,7 +188,7 @@ abstract class BaseSharingTestScenarios extends BaseSharingTest
 	 * it and logs out.
 	 *
 	 * This scenario is similar to
-	 * {@link SharingTestScenariosAbstract::prepareUnsortedPublicAndPrivatePhoto()}
+	 * {@link BaseSharingTestScenarios::prepareUnsortedPublicAndPrivatePhoto()}
 	 * but with an album.
 	 *
 	 * @return void
@@ -521,7 +513,7 @@ abstract class BaseSharingTestScenarios extends BaseSharingTest
 	}
 
 	/**
-	 * Like {@link SharingTestScenariosAbstract::preparePublicAlbumAndPasswordProtectedAlbum},
+	 * Like {@link BaseSharingTestScenarios::preparePublicAlbumAndPasswordProtectedAlbum},
 	 * but additionally the password-protected photo is starred.
 	 *
 	 * @return void
@@ -786,7 +778,7 @@ abstract class BaseSharingTestScenarios extends BaseSharingTest
 	}
 
 	/**
-	 * Like {@link SharingTestScenariosAbstract::preparePublicAlbumAndHiddenAlbum}, but
+	 * Like {@link BaseSharingTestScenarios::preparePublicAlbumAndHiddenAlbum}, but
 	 * additionally the hidden album is also password protected.
 	 *
 	 * @return void
