+++ tests/Feature/Constants/TestConstants.php
@@ -85,9 +85,8 @@ class TestConstants
 	public const CONFIG_MAP_INCLUDE_SUBALBUMS = 'map_include_subalbums';
 	public const CONFIG_PHOTOS_SORTING_COL = 'sorting_photos_col';
 	public const CONFIG_PHOTOS_SORTING_ORDER = 'sorting_photos_order';
-	public const CONFIG_PUBLIC_HIDDEN = 'public_photos_hidden';
 	public const CONFIG_PUBLIC_RECENT = 'public_recent';
-	public const CONFIG_PUBLIC_SEARCH = 'public_search';
+	public const CONFIG_PUBLIC_SEARCH = 'search_public';
 	public const CONFIG_PUBLIC_STARRED = 'public_starred';
 	public const CONFIG_PUBLIC_ON_THIS_DAY = 'public_on_this_day';
 	public const CONFIG_RAW_FORMATS = 'raw_formats';
