+++ tests/Feature/SettingsTest.php
@@ -179,8 +179,8 @@ class SettingsTest extends AbstractTestCase
 	// Route::post('/Settings::setPublicSearch', [Administration\SettingsController::class, 'setPublicSearch']);
 	public function testSetPublicSearch(): void
 	{
-		$this->sendKV('/Settings::setPublicSearch', 'public_search', 'wrong', 422);
-		$this->sendKV('/Settings::setPublicSearch', 'public_search', false);
+		$this->sendKV('/Settings::setPublicSearch', 'search_public', 'wrong', 422);
+		$this->sendKV('/Settings::setPublicSearch', 'search_public', false);
 	}
 
 	// Route::post('/Settings::setCSS', [Administration\SettingsController::class, 'setCSS']);
