+++ tests/Feature/SharingWithNonAdminUserTest.php
@@ -10,7 +10,7 @@
  * @noinspection PhpUnhandledExceptionInspection
  */
 
-namespace Tests\Feature\Base;
+namespace Tests\Feature;
 
 use App\SmartAlbums\OnThisDayAlbum;
 use App\SmartAlbums\PublicAlbum;
@@ -18,6 +18,7 @@ use App\SmartAlbums\RecentAlbum;
 use App\SmartAlbums\StarredAlbum;
 use App\SmartAlbums\UnsortedAlbum;
 use Illuminate\Support\Facades\Auth;
+use Tests\Feature\Base\BaseSharingTestScenarios;
 use Tests\Feature\Constants\TestConstants;
 
 /**
@@ -25,8 +26,13 @@ use Tests\Feature\Constants\TestConstants;
  * non-admin user whose results are independent of the setting for public
  * search.
  */
-abstract class BaseSharingWithNonAdminUser extends BaseSharingTestScenarios
+class SharingWithNonAdminUserTest extends BaseSharingTestScenarios
 {
+	public function setUp(): void
+	{
+		parent::setUp();
+	}
+
 	/**
 	 * {@inheritDoc}
 	 */
@@ -378,5 +384,190 @@ abstract class BaseSharingWithNonAdminUser extends BaseSharingTestScenarios
 	protected function getExpectedDefaultInaccessibleMessage(): string
 	{
 		return TestConstants::EXPECTED_FORBIDDEN_MSG;
+	}
+
+	/**
+	 * Ensures that the user does not the unsorted public photos as covers nor
+	 * inside "Recent", "On This Day" and "Favorites" (as public search is disabled).
+	 * The user can access the public photo nonetheless, but gets
+	 * "403 - Forbidden" for the other.
+	 *
+	 * See {@link BaseSharingTestScenarios::prepareUnsortedPublicAndPrivatePhoto()}
+	 * for description of scenario.
+	 *
+	 * @return void
+	 */
+	public function testUnsortedPublicAndPrivatePhoto(): void
+	{
+		$this->prepareUnsortedPublicAndPrivatePhoto();
+
+		$this->ensurePhotosWereTakenOnThisDay($this->photoID1, $this->photoID2);
+
+		$responseForRoot = $this->root_album_tests->get();
+		$responseForRoot->assertJson($this->generateExpectedRootJson());
+		$arrayUnexpected = $this->generateUnexpectedRootJson();
+		if ($arrayUnexpected !== null) {
+			$responseForRoot->assertJsonMissing($arrayUnexpected);
+		}
+		$responseForRoot->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForRoot->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForUnsorted = $this->albums_tests->get(UnsortedAlbum::ID);
+		$responseForUnsorted->assertJson($this->generateExpectedSmartAlbumJson(false));
+		$responseForUnsorted->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForUnsorted->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForRecent = $this->albums_tests->get(RecentAlbum::ID);
+		$responseForRecent->assertJson($this->generateExpectedSmartAlbumJson(true));
+		$responseForRecent->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForRecent->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForStarred = $this->albums_tests->get(StarredAlbum::ID);
+		$responseForStarred->assertJson($this->generateExpectedSmartAlbumJson(true));
+		$responseForStarred->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForStarred->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForOnThisDay = $this->albums_tests->get(OnThisDayAlbum::ID);
+		$responseForOnThisDay->assertJson($this->generateExpectedSmartAlbumJson(true));
+		$responseForOnThisDay->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForOnThisDay->assertJsonMissing(['id' => $this->photoID2]);
+
+		// Even though the public photo is not searchable and hence does not
+		// show up in the smart albums, it can be fetched directly
+		$this->photos_tests->get($this->photoID1);
+		$this->photos_tests->get($this->photoID2, $this->getExpectedInaccessibleHttpStatusCode());
+	}
+
+	/**
+	 * Ensures that the user does not see any photo, although the first is
+	 * public (but not searchable).
+	 * The first photo is still visible if directly accessed, but the user
+	 * gets a `403 - Forbidden` for the album and the second photo.
+	 *
+	 * See
+	 * {@link BaseSharingTestScenarios::preparePublicAndPrivatePhotoInPrivateAlbum()}
+	 * for description of scenario.
+	 */
+	public function testPublicAndPrivatePhotoInPrivateAlbum(): void
+	{
+		$this->preparePublicAndPrivatePhotoInPrivateAlbum();
+
+		$this->ensurePhotosWereTakenOnThisDay($this->photoID1, $this->photoID2);
+
+		$responseForRoot = $this->root_album_tests->get();
+		$responseForRoot->assertJson($this->generateExpectedRootJson());
+		$arrayUnexpected = $this->generateUnexpectedRootJson();
+		if ($arrayUnexpected !== null) {
+			$responseForRoot->assertJsonMissing($arrayUnexpected);
+		}
+		$responseForRoot->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForRoot->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForUnsorted = $this->albums_tests->get(UnsortedAlbum::ID);
+		$responseForUnsorted->assertJson($this->generateExpectedSmartAlbumJson(false));
+		$responseForUnsorted->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForUnsorted->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForRecent = $this->albums_tests->get(RecentAlbum::ID);
+		$responseForRecent->assertJson($this->generateExpectedSmartAlbumJson(true));
+		$responseForRecent->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForRecent->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForStarred = $this->albums_tests->get(StarredAlbum::ID);
+		$responseForStarred->assertJson($this->generateExpectedSmartAlbumJson(true));
+		$responseForStarred->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForStarred->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForOnThisDay = $this->albums_tests->get(OnThisDayAlbum::ID);
+		$responseForOnThisDay->assertJson($this->generateExpectedSmartAlbumJson(true));
+		$responseForOnThisDay->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForOnThisDay->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForTree = $this->root_album_tests->getTree();
+		$responseForTree->assertJson($this->generateExpectedTreeJson());
+		$responseForTree->assertJsonMissing(['id' => $this->albumID1]);
+		$responseForTree->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForTree->assertJsonMissing(['id' => $this->photoID2]);
+
+		$this->albums_tests->get($this->albumID1, $this->getExpectedInaccessibleHttpStatusCode(), $this->getExpectedDefaultInaccessibleMessage(), TestConstants::EXPECTED_PASSWORD_REQUIRED_MSG);
+		// Even though public search is disabled, the photo is accessible
+		// by its direct link, because it is public.
+		$this->photos_tests->get($this->photoID1);
+		$this->photos_tests->get($this->photoID2, $this->getExpectedInaccessibleHttpStatusCode());
+	}
+
+	public function testPublicUnsortedPhotoAndPhotoInSharedAlbum(): void
+	{
+		$this->preparePublicUnsortedPhotoAndPhotoInSharedAlbum();
+
+		$this->ensurePhotosWereTakenOnThisDay($this->photoID1, $this->photoID2);
+
+		$responseForRoot = $this->root_album_tests->get();
+		$responseForRoot->assertJson($this->generateExpectedRootJson(
+			null,
+			$this->photoID2,
+			null,
+			$this->photoID2,
+			$this->photoID2, [
+				$this->generateExpectedAlbumJson($this->albumID1, TestConstants::ALBUM_TITLE_1, null, $this->photoID2),
+			]
+		));
+		$arrayUnexpected = $this->generateUnexpectedRootJson(
+			null, $this->photoID1, null, $this->photoID2
+		);
+		if ($arrayUnexpected !== null) {
+			$responseForRoot->assertJsonMissing($arrayUnexpected);
+		}
+		$responseForRoot->assertJsonMissing(['id' => $this->photoID1]);
+
+		$responseForUnsorted = $this->albums_tests->get(UnsortedAlbum::ID);
+		$responseForUnsorted->assertJson($this->generateExpectedSmartAlbumJson(false));
+		$responseForUnsorted->assertJsonMissing(['id' => $this->photoID1]);
+		$responseForUnsorted->assertJsonMissing(['id' => $this->photoID2]);
+
+		$responseForRecent = $this->albums_tests->get(RecentAlbum::ID);
+		$responseForRecent->assertJson($this->generateExpectedSmartAlbumJson(
+			true,
+			$this->photoID2, [
+				$this->generateExpectedPhotoJson(TestConstants::SAMPLE_FILE_MONGOLIA_IMAGE, $this->photoID2, $this->albumID1),
+			]
+		));
+		$responseForRecent->assertJsonMissing(['id' => $this->photoID1]);
+
+		$responseForStarred = $this->albums_tests->get(StarredAlbum::ID);
+		$responseForStarred->assertJson($this->generateExpectedSmartAlbumJson(
+			true,
+			$this->photoID2, [
+				$this->generateExpectedPhotoJson(TestConstants::SAMPLE_FILE_MONGOLIA_IMAGE, $this->photoID2, $this->albumID1),
+			]
+		));
+		$responseForStarred->assertJsonMissing(['id' => $this->photoID1]);
+
+		$responseForOnThisDay = $this->albums_tests->get(OnThisDayAlbum::ID);
+		$responseForOnThisDay->assertJson($this->generateExpectedSmartAlbumJson(true,
+			$this->photoID2, [
+				$this->generateExpectedPhotoJson(TestConstants::SAMPLE_FILE_MONGOLIA_IMAGE, $this->photoID2, $this->albumID1),
+			]
+		));
+		$responseForOnThisDay->assertJsonMissing(['id' => $this->photoID1]);
+
+		$responseForTree = $this->root_album_tests->getTree();
+		$responseForTree->assertJson($this->generateExpectedTreeJson([
+			$this->generateExpectedAlbumJson($this->albumID1, TestConstants::ALBUM_TITLE_1, null, $this->photoID2),
+		]));
+
+		$responseForAlbum = $this->albums_tests->get($this->albumID1);
+		$responseForAlbum->assertJson([
+			'id' => $this->albumID1,
+			'title' => TestConstants::ALBUM_TITLE_1,
+			'policy' => ['is_public' => false],
+			'thumb' => $this->generateExpectedThumbJson($this->photoID2),
+			'photos' => [
+				$this->generateExpectedPhotoJson(TestConstants::SAMPLE_FILE_MONGOLIA_IMAGE, $this->photoID2, $this->albumID1),
+			],
+		]);
+
+		$this->photos_tests->get($this->photoID1);
+		$this->photos_tests->get($this->photoID2);
 	}
 }
